extract movement
