+++ crates/client/src/app/input/mod.rs
@@ -9,3 +9,4 @@ pub mod multi_action;
 pub mod npc_interaction;
 pub mod splash;
 pub mod world_build;
+pub mod world_movement;
